+++ src/app/(content-layout)/[teamid]/tasklist/page.tsx
@@ -5,22 +5,12 @@ import nextIcon from '@/../public/icons/next-arrow-icon.svg';
 import calendar from '@/../public/icons/calendar.svg';
 import Image from 'next/image';
 import { addDays, format, subDays } from 'date-fns';
-
 import { useState } from 'react';
-import clsx from 'clsx';
-import { TaskTest1, TaskTest2, TaskTest3 } from './_tasklist/mock-data-task-list-page';
-import { TaskListApiResponse } from './_tasklist/types/task-list-page-type';
-import ListItems from './_tasklist/components/ListItems';
 import Button from '@/components/common/Button';
-
-const MOCK_DATA = [TaskTest1, TaskTest2, TaskTest3];
+import DateWiseTaskList from './_tasklist/components/DateWiseTaskList';
 
 export default function Page() {
   const [currentDate, setCurrentDate] = useState(new Date());
-  const formattedDateToKorean = format(currentDate, 'M월 dd일');
-
-  const [currentTask, setCurrentTask] = useState(MOCK_DATA[0]);
-  const [currentTaskItem, setCurrentTaskItem] = useState(currentTask.tasks);
 
   const handleClickChangeDayIcon = (value: string) => {
     if (value === 'prev') {
@@ -32,11 +22,6 @@ export default function Page() {
 
   const handleClickCalendarPopUp = () => {
     // 캘린더 띄움
-  };
-
-  const handleClickChangeCurrentTask = (task: TaskListApiResponse) => {
-    setCurrentTask(task);
-    setCurrentTaskItem(() => task.tasks);
   };
 
   const handleClickCreateTask = () => {
@@ -54,7 +39,7 @@ export default function Page() {
         <p className="text-lg-bold md:text-xl-bold">할일</p>
         <div className="flex justify-between">
           <div className="flex items-center gap-3">
-            <p className="text-lg-md">{formattedDateToKorean}</p>
+            <p className="text-lg-md">{format(currentDate, 'M월 dd일')}</p>
             <div className="flex gap-1">
               <button onClick={() => handleClickChangeDayIcon('prev')}>
                 <Image src={prevIcon} width={16} height={16} alt="<" />
@@ -63,37 +48,15 @@ export default function Page() {
                 <Image src={nextIcon} width={16} height={16} alt=">" />
               </button>
             </div>
-
             <button onClick={handleClickCalendarPopUp}>
               <Image src={calendar} width={24} height={24} alt=">" />
             </button>
           </div>
-
           <div onClick={handleClickCreateTask} className="text-primary text-md-rg">
             + 새로운 목록 추가하기
           </div>
         </div>
-        <div className="flex flex-col gap-4">
-          <div className="flex gap-3">
-            {MOCK_DATA.map((task) => {
-              return (
-                <p
-                  key={task.id}
-                  onClick={() => handleClickChangeCurrentTask(task)}
-                  className={clsx(
-                    'text-md-md cursor-pointer',
-                    task === currentTask
-                      ? 'text-gray-200 underline underline-offset-6'
-                      : 'text-gray-500'
-                  )}
-                >
-                  {task.name}
-                </p>
-              );
-            })}
-          </div>
-          <ListItems ListItem={currentTaskItem} />
-        </div>
+        <DateWiseTaskList date={currentDate} />
       </div>
       <Button
         className="absolute right-6 -bottom-20"
